+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/app.module.ts
@@ -1,14 +1,17 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-
+import { CommonModule } from '@angular/common';
 import { AppComponent } from './app.component';
+import { ComponentsModule } from './components/components.module';
 
 @NgModule({
   declarations: [
     AppComponent
   ],
   imports: [
-    BrowserModule
+    BrowserModule,
+    ComponentsModule,
+    CommonModule
   ],
   providers: [],
   bootstrap: [AppComponent]