+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/header/header.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { HeaderComponent } from './header.component';
+import { SelectListModule } from '../select-list/select-list.module';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    SelectListModule
+  ],
+  declarations: [HeaderComponent],
+  exports: [
+    HeaderComponent
+  ]
+})
+export class HeaderModule { }