+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/header/header.ts
@@ -1,0 +1,5 @@
+import { SelectListOptions } from '../select-list';
+
+export interface HeaderOptions {
+  selectListOptions?: SelectListOptions;
+}