+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/index.ts
@@ -1,0 +1,5 @@
+export {Subscriptions } from './subscriptions';
+export * from './select-list';
+export * from './panel';
+export * from './header';
+