+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/index.ts
@@ -1,0 +1,3 @@
+export {PanelComponent} from './panel.component';
+export {PanelOptions} from './panel';
+export {PanelEventTypes, PanelEvent} from './panel-event';