+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/panel-event.ts
@@ -1,0 +1,11 @@
+
+export enum PanelEventTypes {
+  DRAG_START,
+  DRAG_OVER,
+  DRAG_END
+}
+
+export interface PanelEvent {
+  type: PanelEventTypes,
+  data: any;
+}