+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/panel-header/index.ts
@@ -1,0 +1,3 @@
+export {PanelHeaderComponent} from './panel-header.component';
+export {PanelHeaderOptions} from './panel-header';
+export {PanelHeaderModule} from './panel-header.module';