+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/panel-header/panel-header.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { PanelHeaderComponent } from './panel-header.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [PanelHeaderComponent],
+  exports: [
+    PanelHeaderComponent
+  ]
+})
+export class PanelHeaderModule { }