+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/panel.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { PanelComponent } from './panel.component';
+import { PanelHeaderModule } from './panel-header/panel-header.module';
+
+@NgModule({
+  imports: [
+    CommonModule,
+    PanelHeaderModule
+  ],
+  declarations: [PanelComponent],
+  exports: [
+    PanelComponent
+  ]
+})
+export class PanelModule { }