+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/panel/panel.ts
@@ -1,0 +1,11 @@
+export interface PanelOptions {
+  seq: number,
+  url: string;
+  width?: number;
+  height?: number;
+  top?: number;
+  left?: number;
+  title?: string;
+  isActive?: boolean;
+
+}