+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/select-list/select-list.module.ts
@@ -1,0 +1,15 @@
+import { NgModule} from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SelectListComponent } from './select-list.component';
+
+@NgModule({
+  imports: [
+    CommonModule
+  ],
+  declarations: [SelectListComponent],
+  exports: [
+    SelectListComponent
+  ]
+
+})
+export class SelectListModule { }