+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/components/select-list/select-list.service.spec.ts
@@ -1,0 +1,15 @@
+import { TestBed, inject } from '@angular/core/testing';
+
+import { SelectListService } from './select-list.service';
+
+describe('SelectListService', () => {
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      providers: [SelectListService]
+    });
+  });
+
+  it('should be created', inject([SelectListService], (service: SelectListService) => {
+    expect(service).toBeTruthy();
+  }));
+});