+++ fiddles/angular2-cli/fiddle-0012-PanelReOrdering/src/app/store/selectors.ts
@@ -1,0 +1,16 @@
+export const framesModel = () => {
+  return state => state
+    .map(([frames, filter]) => {
+      return {
+        total: frames.length,
+        frames: frames.filter(filter),
+        active: frames.filter(frames => frames.isActive).length
+      }
+    })
+};
+
+export const activeFrames = () => {
+  return state => state
+    .map(s => s.frame)
+    .distinctUntilChanged();
+};